refactor: extract ContextMenuItems component
for reuse e.g. in node toolbar more menu
