+++ src/web/common/components/ContextMenu/ContextMenuItems.tsx
@@ -1,0 +1,54 @@
+import { AddNodeMenuItem } from "@/web/common/components/ContextMenu/AddNodeMenuItem";
+import { AlsoShowNodeAndNeighborsMenuItem } from "@/web/common/components/ContextMenu/AlsoShowNodeAndNeighborsMenuItem";
+import { ChangeEdgeTypeMenuItem } from "@/web/common/components/ContextMenu/ChangeEdgeTypeMenuItem";
+import { ChangeNodeTypeMenuItem } from "@/web/common/components/ContextMenu/ChangeNodeTypeMenuItem";
+import { ContextMenuItem } from "@/web/common/components/ContextMenu/CloseOnClickMenuItem";
+import { CopyLinkToPartMenuItem } from "@/web/common/components/ContextMenu/CopyLinkToPart";
+import { DeleteEdgeMenuItem } from "@/web/common/components/ContextMenu/DeleteEdgeMenuItem";
+import { DeleteNodeMenuItem } from "@/web/common/components/ContextMenu/DeleteNodeMenuItem";
+import { HideMenuItem } from "@/web/common/components/ContextMenu/HideMenuItem";
+import { OnlyShowNodeAndNeighborsMenuItem } from "@/web/common/components/ContextMenu/OnlyShowNodeAndNeighborsMenuItem";
+import { ViewContextInDiagramMenuItem } from "@/web/common/components/ContextMenu/ViewContextInDiagramMenuItem";
+import { ViewDetailsMenuItem } from "@/web/common/components/ContextMenu/ViewDetailsMenuItem";
+import { ViewTableMenuItem } from "@/web/common/components/ContextMenu/ViewTableMenuItem";
+import { Context } from "@/web/common/store/contextMenuStore";
+
+/**
+ * Extracted to a separate component so that we can reuse for both the regular context menu and for
+ * the node's more menu.
+ */
+export const ContextMenuItems = ({
+  node: contextNode,
+  edge: contextEdge,
+  isOpen,
+}: Context & { isOpen: boolean }) => {
+  const contextPart = contextNode ?? contextEdge;
+
+  // create these based on what's set in the context
+  const menuItems = [
+    // view actions (so that this functionality is still available if indicators are hidden)
+    contextPart && <ViewDetailsMenuItem graphPart={contextPart} key={1} />,
+    contextNode?.type === "problem" && <ViewTableMenuItem node={contextNode} key={2} />,
+    contextPart && <ViewContextInDiagramMenuItem graphPart={contextPart} key={3} />,
+
+    // CRUD actions
+    contextPart === undefined && <AddNodeMenuItem parentMenuOpen={isOpen} key={9} />,
+    contextNode && <ChangeNodeTypeMenuItem node={contextNode} parentMenuOpen={isOpen} key={7} />,
+    contextEdge && <ChangeEdgeTypeMenuItem edge={contextEdge} parentMenuOpen={isOpen} key={8} />,
+    contextNode && <DeleteNodeMenuItem node={contextNode} key={5} />,
+    contextEdge && <DeleteEdgeMenuItem edge={contextEdge} key={6} />,
+
+    // show/hide actions
+    contextNode && <AlsoShowNodeAndNeighborsMenuItem node={contextNode} key={11} />,
+    contextNode && <OnlyShowNodeAndNeighborsMenuItem node={contextNode} key={12} />,
+    contextNode && <HideMenuItem node={contextNode} key={13} />,
+
+    // misc? actions
+    contextPart && <CopyLinkToPartMenuItem graphPart={contextPart} key={14} />,
+
+    // ensure there's never an empty context menu; that shows an empty bubble and feels awkward
+    <ContextMenuItem key={10}>Cancel</ContextMenuItem>,
+  ].filter((item): item is JSX.Element => !!item);
+
+  return <>{menuItems}</>;
+};
